Further edits of Linux_basics.
Added a couple of files too, one is a markdown cheatsheet from
github, anotherone is a gulpfile with a few things I've been
reading. And systemUtils.txt that is gonna become a html file
to append to linux files. That one is about utilities like
ps for viewing processes, htop, top, etc.